% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puccRawDataDumpCleaningRmdScript_utils.R
\name{clean_partitioned_pucc_data_dump_file}
\alias{clean_partitioned_pucc_data_dump_file}
\title{Cleaning function for PUC data dump}
\usage{
clean_partitioned_pucc_data_dump_file(
  df,
  input_file_name,
  clean_output_folder_rel_path
)
}
\description{
Cleaning function for PUC data dump
}
