% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puccRawDataDumpPartitioningRmdScript_utils.R
\name{combine_partitioned_files}
\alias{combine_partitioned_files}
\title{Combine partitioned files of a given table into a single master dataframe.}
\usage{
combine_partitioned_files(rel_path_to_directory, keyword)
}
\arguments{
\item{rel_path_to_directory}{relative path to the directory containing files to be combined}

\item{keyword}{a keyword that uniquely identifies the type of table. The function will look for this
keyword in the file name.}
}
\value{
a data.table that is created by horizontally row-binding all the files (that contains the
\code{keyword}) in the the relative directory path.
}
\description{
This function detects the files to be combined (in a given directory that is supplied as an argument to this function)
based on the unique keyword. This keyword is supplied as an argument to this function. It then combines
those individual files into a master data frame. It leverages the power of data.table package in R.
}
\examples{
combine_partitioned_files("rel/path/to/directory/containing/filesWithKeyWord", keyword = "foo")

}
