% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puccRawDataDumpPartitioningRmdScript_utils.R
\name{combine_partitioned_files_v2}
\alias{combine_partitioned_files_v2}
\title{Combine partitioned files of a given table into a single master dataframe (v2).}
\usage{
combine_partitioned_files_v2(
  rel_path_to_directory,
  keyword,
  output_dir_rel_path,
  output_file_name
)
}
\description{
This function detects the files to be combined (in a given directory that is supplied as an argument to this function)
based on the unique keyword. This keyword is supplied as an argument to this function. It then combines
those individual files into a master data frame. It leverages the power of data.table package in R.
V2 allows for a lot of files and can handle more than "x" GB of data.
}
