% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puccRawDataDumpCleaningRmdScript_utils.R
\name{convert_character_col_case}
\alias{convert_character_col_case}
\title{convert character columns into lower/upper case}
\usage{
convert_character_col_case(df, case = "lower")
}
\arguments{
\item{df}{data frame whose character columns need to be converted}

\item{case}{this can take one of two values: "lower": for converting to lower case
or "upper" for converting to upper case.}
}
\value{
returns the \code{df} but with character columns now converted to upper/lower
        case (as mentioned in the \code{case} parameter.
}
\description{
This function converts all columns of type character and converts them
into lower or upper case.
}
\examples{
convert_character_col_case(df, "upper") # for converting to upper case.
convert_character_col_case(df, "lower") # for converting to lower case.

}
