% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puccRawDataDumpCleaningRmdScript_utils.R
\name{find_duplicates_in_remote_df_cols}
\alias{find_duplicates_in_remote_df_cols}
\title{Find column wise duplicates for a given table.}
\usage{
find_duplicates_in_remote_df_cols(df, output_folder_df_col_wise_dup_count)
}
\description{
This function takes as its input a dataframe and exports in the given output folder
one excel file each for each column of that dataframe. This exported excel file
contains the list of duplicate values in that column. Works for all types of
data, including remote databaes. In fact this was specifically written for
a postgres database.
}
