% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puccRawDataDumpCleaningRmdScript_utils.R
\name{find_primary_keys}
\alias{find_primary_keys}
\title{identify primary keys in a dataset}
\usage{
find_primary_keys(df)
}
\arguments{
\item{df}{data frame whose primary keys need to be figured out}
}
\value{
a character vector of primary keys (if any), else return a message
        that no primary keys are available.
}
\description{
This function takes a dataset as its input and returns the primary
key columns (if any) for that dataset.
}
\examples{
find_primary_keys(mtcars)

}
