% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puccRawDataDumpCleaningRmdScript_utils.R
\name{find_uniq_or_dup_vals_in_remote_df_cols}
\alias{find_uniq_or_dup_vals_in_remote_df_cols}
\title{find unique or duplicate values in cols of a given remote dataframe.
This adds to the "find_duplicates_in_remote_df_cols" function by not allowing
for an operation type specification e.g. if we want to find unique values,
we enter the keyword: "unique", else we enter the keyword "duplicate".}
\usage{
find_uniq_or_dup_vals_in_remote_df_cols(
  df,
  output_folder_df_col_wise_dup_or_unq_count,
  type
)
}
\description{
find unique or duplicate values in cols of a given remote dataframe.
This adds to the "find_duplicates_in_remote_df_cols" function by not allowing
for an operation type specification e.g. if we want to find unique values,
we enter the keyword: "unique", else we enter the keyword "duplicate".
}
