% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puccRawDataDumpCleaningRmdScript_utils.R
\name{find_unique_values_by_col}
\alias{find_unique_values_by_col}
\title{get unique values for all columns for a given data frame}
\usage{
find_unique_values_by_col(df)
}
\arguments{
\item{df}{the dataframe for which the column wise unique values are needed.}
}
\value{
a list whose xth element contain the unique values for the xth column
        of the \code{df}.
}
\description{
For a given data frame, find the unique values for each column. Return
a list whose each element contains the unique values of each corresponding
column of the dataset. For example, the first element of the returned
list would contain the unique values of the first column of the dataset,
and so on.
}
\examples{
find_unique_values_by_col(df = mtcars)


}
