% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puccRawDataDumpPartitioningRmdScript_utils.R
\name{partition_data_pucc_specific}
\alias{partition_data_pucc_specific}
\title{Partition all .csv Excel files in a given directory for the PUCC data dump}
\usage{
partition_data_pucc_specific(
  directory_rel_path,
  test_mode = "yes",
  rename_files = "no",
  partition_at = 1e+06,
  test_df = mtcars,
  pucc_col_type_name_specify = FALSE
)
}
\arguments{
\item{directory_rel_path}{relative path to the directory containing the .csv files}

\item{test_mode}{if trying out the function, before the real work, set this to "yes". This tests the
partitioning on a test dataset. The test dataset can be specified in \code{test_df}
argument.}

\item{rename_files}{if this is set to "yes", the files in the directory path specified will be renamed.
This is again a, pucc_specific argument. Also, as of now, it is inactive, and
defaults to "no". So, please name your master files appropriately before partitioning.}

\item{partition_at}{the row number at which, post which a partition should be created. Default to 1e+06}

\item{test_df}{this is the dataframe used when \code{test_mode} = "yes"}

\item{pucc_col_type_name_specify}{if this is TRUE, the pucc data columns will be apprpriately named. Also,
all column types for pucc would be set. Note that this only applies
to the files received from the transport department.}
}
\value{
returns a list whose first element is a list of lists. In this list of lists, the first layer represents
                the master files, and the second layer contains the partitions of these files
                (again stored as a list). So, the returned object is a 3 layer deep list, which
                is navaigated using x[[i]][[j]][[k]], syntax.
}
\description{
This function partitions each .csv file (in a given directory). The total number
of partitions depends on the threshold at which we would like to partition the
files in the directory. If "pucc_col_type_name_specify" = TRUE, then this function
names the columns for the pucc files. As of now, this feature is only work for
the files that we received from the transport department. Later on, I will write
a more general function. To simply partition the files (without setting column names,
column types), set "pucc_col_type_name_specify" = FALSE.
}
\examples{
partition_data_pucc_specific("./path/to/data/directory", test_mode = "no", partition_at = 500000)
partition_data_pucc_specific("./path/to/data/directory", test_mode = "yes", test_df = mtcars)

}
