% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallTasksRmdScript_utils.R
\name{plot_bar_transp_dept_format_data}
\alias{plot_bar_transp_dept_format_data}
\title{This function plots the bar graphs for data that for which counts for
various categories is available.}
\usage{
plot_bar_transp_dept_format_data(
  df,
  map_x,
  map_y,
  fill,
  y_axis_ticks,
  xlab,
  ylab,
  title
)
}
\arguments{
\item{df}{data frame that will be used to plot the data}

\item{map_x}{the column name (in quotes) that will be mapped to the x axis.}

\item{map_y}{the column name (in quotes) that will be mapped to the y axis.}

\item{fill}{the color with which the bar graphs will be filled in.
This should be a character vector of length 1.}

\item{y_axis_ticks}{the spacing of tick marks on the y axis.
this should be a numeric vector containing
points at which tick marks need to be placed.}

\item{xlab}{x axis label. This should be a character vector of length 1.}

\item{ylab}{y axis label. This should be a character vector of length 1.}

\item{title}{Graph Title. This should be a character vector of length 1.}
}
\value{
a ggplot object.
}
\description{
This function plots the bar graphs for data that for which counts for
various categories is available.
}
\examples{
plot_bar_transp_dept_format_data(dataset, "col_a", "count_for_col_a", "midnightblue", "col_a", "col_a_count",
"Graph Title")

}
