% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puccSurveyReportRmd_utils.R
\name{plot_bargraph}
\alias{plot_bargraph}
\title{Plot custom bar graph for a given discrete factor column}
\usage{
plot_bargraph(
  df,
  col_to_plot,
  type_of_graph = "percentage",
  xlab = " ",
  ylab = " ",
  title = " ",
  facet = FALSE,
  facet_by = " "
)
}
\arguments{
\item{df}{dataframe used to generate the bar graph}

\item{col_to_plot}{name (character string of length 1) of the  column from
\code{df} whose distribution is to be plotted.}

\item{type_of_graph}{this can take one of two character values: "count" or
"percentage"}

\item{xlab}{label for the x-axis}

\item{ylab}{label for the y-axis}

\item{title}{title for the graph}

\item{facet}{if set to \code{TRUE}, a facetted graph will be produced. By default
the value of this parameter is set as \code{FALSE}.}

\item{facet_by}{a character string of length 1, specifying the name of the variable
/column by which we want to facet the graph}
}
\value{
a ggplot object containing the graph
}
\description{
This function plots a custom bar graph for a given factor column.
}
\examples{

plot_bargraph(df = mtcars, col_to_plot = "gear", type_of_graph = "percentage",
xlab = "gear", ylab = "percentage",
title = "distribution of gears for mtcars dataset", facet = FALSE)


}
