% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puccRawDataDumpCleaningRmdScript_utils.R
\name{rename_cols_pucc_data_dump}
\alias{rename_cols_pucc_data_dump}
\title{rename columns for pucc data dump tables}
\usage{
rename_cols_pucc_data_dump(
  rel_path_input_folder,
  rel_path_output_folder,
  t1_new_col_names,
  t2_new_col_names,
  t3_new_col_names,
  t4_new_col_names,
  t5_new_col_names
)
}
\description{
This function renames the columns of all given csv files (which should belong to
one of the five categories of table types in the pucc data dump) in a given input folder.
It takes in as its inputs the path to the input folder, path to the output folder,
5 character vectors (one for each table) containing the new column names for that table.
For all the files in the given input directory, it automatically detects the "table
type" from the file name (which follows a specific naming scheme) and then reassigns
the column names. Once the names are reassigned, it writes and exports these new
csv files (with new column names) to the specified output folder.
}
