% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puccRawDataDumpCleaningRmdScript_utils.R
\name{replace_col_values}
\alias{replace_col_values}
\title{Clean columns based on pre-specified regular expressions}
\usage{
replace_col_values(df, fuel_type_col_name, from_reg_exps, to_reg_exps)
}
\arguments{
\item{df}{data frame inside of which the column values need to be replaced.}

\item{fuel_type_col_name}{name of the column (within which replacements will happen).
This should be a character vector of length 1.}

\item{from_reg_exps}{a character vector of "from" regular expressions.}

\item{to_reg_exps}{a character vector (corresponding to \code{from_reg_exps})
of "to" regular expressions.}
}
\value{
returns \code{df} but with column values changed as per the \code{to_reg_exps}
}
\description{
This function is built specifically for cleaning a column based on regular
expressions. This takes in a data frame, a column name from that data frame,
a character vector of "from" regular expressions, a character vector of "to"
regular expressions. In the column specified, the function detects the entries
that match the "from" regular expression pattern and convert them to the "to"
regular expression.
}
\examples{
replace_col_values(datasets::CO2, "Type", "Q", "q")

}
