% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puccSurveyReportRmd_utils.R
\name{summarise_df_by_colname}
\alias{summarise_df_by_colname}
\title{Generate a summary table given a grouping variable}
\usage{
summarise_df_by_colname(df, col_to_grp_by)
}
\arguments{
\item{df}{dataframe to be summarised given \code{col_to_grp_by}}

\item{col_to_grp_by}{a character vector of column name(s) to group by}

\item{arrange_by}{a character vector (length 1) of column name to arrange by once grouping
is done in \code{col_to_grp_by}}
}
\value{
a summarised dataframe grouped by \code{col_to_grp_by} containing counts
        and percentages
}
\description{
Given a dataframe, generate a summary table, given a column name by which
to group_by. This table contains both percentages and counts.
}
\examples{
summarise_df_by_colname(df = mtcars, col_to_grp_by = "gear")
summarise_df_by_colname(df = mtcars, col_to_grp_by = c("gear", "cyl"))

}
