% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puccSurveyReportRmdScript_utils.R
\name{time_hms_formatter}
\alias{time_hms_formatter}
\title{time hms formatter for creating labels for \code{time_only} values in graphs}
\usage{
time_hms_formatter(num_seconds_since_12am)
}
\arguments{
\item{num_seconds_since_12am}{this takes in a time value in seconds (with the reference
starting point being 00:00:00) in seconds. So,
for example 1 am in the morning would be equal
to 3600 seconds. So, entering "3600" in
\code{num_seconds_since_12am} will output
"01:00:00" which is 1 am in the morning. The outputs
will be in 24 hr time format.}
}
\value{
a string representing time in "hh:mm:ss" format
}
\description{
Generates custom labels for \code{time_only} column. Once generated these can
be plotted on a graph.
}
\examples{

time_hms_formatter(3600) # This outputs: "01:00:00" which is 3600 seconds past from
"00:00:00"

time_hms_formatter(7200) # This outputs: "02:00:00" which is 7200 seconds past from
"00:00:00"


}
